update: promotion product active

+++ Webapp_Shop_Ecommerce/src/main/java/com/example/webapp_shop_ecommerce/dto/response/productdetails/ProductDetailsBillResponse.java
@@ -23,7 +23,8 @@ public class ProductDetailsBillResponse {
     private String status;
     private SizeResponse size;
     private ColorResponse color;
-    private  Float weight;
+    private  Integer weight;
+
 
     private ProductBillResponse product;
